Makefile fixes Simpler upload process Handle upload and rename events

+++ src/opt/app/commands/move_uploaded.py
@@ -3,18 +3,32 @@ import shutil
 import os
 import config
 import glob
+import fnmatch
+
+
+_filters = config.APP_LANDING_TEMP_PATTERNS.split(',')
 
 
 def _move_file_to_upload_directory(path):
+
+    for _f in _filters:
+        if fnmatch.fnmatch(path, _f):
+            log.info("Skipping file that matches temp pattern: {}".format(_f), path=path)
+            return
+
     destination = os.path.join(config.APP_LANDING_UPLOAD_DIR, os.path.basename(path))
-    log.info("Moving {} to {}".format(path, destination))
-    shutil.move(path, destination)
+    if os.path.exists(path):
+        log.info("Moving {} to {}".format(path, destination))
+        shutil.move(path, destination)
+    else:
+        log.warn("{} doesn't exist, maybe it has been moved?".format(path), path=path)
 
 
 def move_uploaded(message):
-    if 'event' in message and message['event'] == 'sftp_write_file':
-        absolute_path = os.path.join(config.APP_LANDING_DIR, message['path'])
-        _move_file_to_upload_directory(absolute_path)
+    if 'event' in message:
+        if message['event'] in ['sftp_write_file', 'sftp_rename']:
+            absolute_path = os.path.join(config.APP_LANDING_DIR, message['path'].lstrip('/'))
+            _move_file_to_upload_directory(absolute_path)
 
 
 def move_existing():
